ActivityDrivenModel: Generation of reluctances
ActivityDrivenModel::generate_agents_from_power_law(), now also
generates the reluctances from a truncated normal distribution.
Optionally they can be correlated with the activities.

TODO: parsing of the config file to control these options

+++ include/util/math.hpp
@@ -162,7 +162,7 @@ public:
     template<typename Generator>
     ScalarT operator()( Generator & gen )
     {
-        for( int i = 0; i < max_tries; i++ )
+        for( size_t i = 0; i < max_tries; i++ )
         {
             auto sample = normal_dist( gen );
             if( sample > eps )
